fixup: inline private functions.
