+++ hub/state/config.go
@@ -3,14 +3,9 @@ package state
 import (
 	"encoding/json"
 	"io"
-	"os"
 	"path/filepath"
 
-	"github.com/hashicorp/go-multierror"
-	"golang.org/x/xerrors"
-
 	"github.com/greenplum-db/gpupgrade/greenplum"
-	"github.com/greenplum-db/gpupgrade/utils"
 )
 
 type InitializeConfig struct {
@@ -51,45 +46,4 @@ func (c *Config) Save(w io.Writer) error {
 	enc := json.NewEncoder(w)
 	enc.SetIndent("", "  ")
 	return enc.Encode(c)
-}
-
-// Save persists the hub's configuration to disk.
-func save(stateDir string, config *Config) (err error) {
-	// TODO: Switch to an atomic implementation like renameio. Consider what
-	// happens if config.Save() panics: we'll have truncated the file
-	// on disk and the hub will be unable to recover. For now, since we normally
-	// only save the configuration during initialize and any configuration
-	// errors could be fixed by reinitializing, the risk seems small.
-	file, err := utils.System.Create(GetConfigFilepath(stateDir))
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if cerr := file.Close(); cerr != nil {
-			cerr = xerrors.Errorf("closing hub configuration: %w", cerr)
-			err = multierror.Append(err, cerr).ErrorOrNil()
-		}
-	}()
-
-	err = config.Save(file)
-	if err != nil {
-		return xerrors.Errorf("saving hub configuration: %w", err)
-	}
-
-	return nil
-}
-
-func loadConfig(conf *Config, path string) error {
-	file, err := os.Open(path)
-	if err != nil {
-		return xerrors.Errorf("opening configuration file: %w", err)
-	}
-	defer file.Close()
-
-	err = conf.Load(file)
-	if err != nil {
-		return xerrors.Errorf("reading configuration file: %w", err)
-	}
-
-	return nil
 }
